Had to make one change to the model. Had to change the data type of tinyint to int because it mapped to the project as a byte c# data type. then I worked on the maitenance form and got that working and looking good

+++ ThemePark/ThemePark/Models/Ticket.cs
@@ -27,7 +27,7 @@ namespace ThemePark
         [Column(TypeName = "date")]
         public DateTime? DateOfPurchase { get; set; }
 
-        public byte? TicketCode { get; set; }
+        public int? TicketCode { get; set; }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<ADMITTED_BY> ADMITTED_BY { get; set; }
